requirements.txt
langchain>=0.3.21
langchain-openai>=0.3.11
chromadb>=0.4.24
sentence-transformers>=2.5.1
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.4
sqlalchemy>=2.0.28
python-multipart>=0.0.9 
faiss-cpu>=1.9.0
